Ironing out the bugs
